% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_es_new_test.R
\name{cal_es_new_test}
\alias{cal_es_new_test}
\title{cal nes test}
\usage{
cal_es_new_test(dt)
}
\arguments{
\item{dt}{dataframe}

\item{type}{"neo" or "not neo"}
}
\value{

}
\description{
cal nes test
}
