% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_nes_new_test.R
\name{cal_nes_new_test}
\alias{cal_nes_new_test}
\title{cal nes test}
\usage{
cal_nes_new_test(dt, sample_counts = 1000, need_p = TRUE)
}
\arguments{
\item{dt}{dataframe}
}
\value{
es_p
}
\description{
cal nes test
}
