% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_p_and_normalized_simulation.R
\name{cal_nes_simulation}
\alias{cal_nes_simulation}
\title{calculate p and nes for simulation}
\usage{
cal_nes_simulation(es, neo_list, test, cal_type, type, cancer_type = FALSE)
}
\arguments{
\item{es}{Es.}

\item{neo_list}{Neoantigen list.}

\item{test}{Mutation file.}
}
\value{
es,nes,p
}
\description{
calculate p and nes for simulation
}
