% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_p_and_normalized.R
\name{cal_p_and_normalized}
\alias{cal_p_and_normalized}
\title{calculate p value and normalized enrichment score for neoantigens}
\usage{
cal_p_and_normalized(es, neo_list, mutation_dt, cal_type, type, trim)
}
\arguments{
\item{es}{Enrichment score}

\item{neo_list}{Neoantigens list}

\item{mutation_dt}{Mutations dataframe}
}
\value{
enrichment score, enrichment score, p value
}
\description{
calculate p value and normalized enrichment score for neoantigens
}
