% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_es.R
\name{cales}
\alias{cales}
\title{calculate enrichment score for neoantigens}
\usage{
cales(mutation_dt, neoantigen_list, type)
}
\arguments{
\item{mutation_dt}{Mutations dataframe}

\item{neoantigen_list}{Neoantigens list}
}
\value{
enrichment score
}
\description{
calculate enrichment score for neoantigens
}
