% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cales_simulation.R
\name{cales_simulation}
\alias{cales_simulation}
\title{calculate es for simulation}
\usage{
cales_simulation(test, neoantigen_list, cal_type, type)
}
\arguments{
\item{test}{Mutation file.}

\item{neoantigen_list}{Neoantigen list.}
}
\value{
es
}
\description{
calculate es for simulation
}
