% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_nes_combine.R
\name{cales_t_combine}
\alias{cales_t_combine}
\title{calculate neoantigens NES for a sample, combing ccf and exp}
\usage{
cales_t_combine(
  data,
  barcode,
  calp = FALSE,
  mhc_type = "I",
  IC50_threshold = 500,
  Rank_threshold = 10,
  type = "I",
  trim,
  DAI,
  DAI_threshold,
  sample_counts
)
}
\arguments{
\item{data}{Mutation file,folowing column are possible needed: MT_mean,exp,sample,chromosome,position,\%Rank_best_perL,ccf_cn_assume.}

\item{barcode}{Tumor sample barcode.}

\item{calp}{Bool value, calculate p value or not.}

\item{mhc_type}{Should be "I" or "II" ,calculate enrichmrnt score for MHC-I or MHC-II.}

\item{IC50_threshold}{Mutations below this threshold are considered as MHC-I neoantigens.}

\item{Rank_threshold}{Mutations below this threshold are considered as MHC-II neoantigens.}
}
\value{
es,nes,p
}
\description{
calculate neoantigens NES for a sample, combing ccf and exp
}
