% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancer_type_code}
\alias{cancer_type_code}
\title{Cancer type code and TCGA Study Abbreviations}
\format{
A data frame with 830 rows and 3 variables:
\describe{
  \item{cancer_code}{two characters, representing 6 and 7 position of TCGA sample barcode}
  \item{cancer_type_full_name}{full name of TCGA cancer type}
  \item{cancer_type}{TCGA cancer type abbreviations}
}
}
\source{
\url{https://gdc.cancer.gov/resources-tcga-users/tcga-code-tables/tissue-source-site-codes}

\url{https://gdc.cancer.gov/resources-tcga-users/tcga-code-tables/tcga-study-abbreviations}
}
\usage{
cancer_type_code
}
\description{
A dataset containing the relationship between Cancer type code and TCGA Study Abbreviations
}
\keyword{datasets}
