% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mutation_type.R
\name{get_mutation_type}
\alias{get_mutation_type}
\title{get mutation type (SBS96) for mutation data}
\usage{
get_mutation_type(mutation)
}
\arguments{
\item{mutation}{mutation data, at least contains following columns: position, chromosome}
}
\value{
mutation data with mutation type
}
\description{
get mutation type (SBS96) for mutation data
}
