% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cancer_type.R
\name{getcancer_type}
\alias{getcancer_type}
\title{get cancer type from tumor sample barcode}
\usage{
getcancer_type(tumor_barcode, cores, par = FALSE)
}
\arguments{
\item{tumor_barcode}{TCGA tumor sample barcode}

\item{cores}{How many cores to use if parallelization is used}

\item{par}{bool value, use parallelization or not}
}
\value{
cancer type
}
\description{
get cancer type from tumor sample barcode
}
\examples{
function("TCGA-OR-A5J1",par=FALSE)
}
