% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_gm_mean.R
\name{gm_mean}
\alias{gm_mean}
\title{Calculate geometric mean}
\usage{
gm_mean(x, na.rm = TRUE, zero.propagate = FALSE)
}
\arguments{
\item{x}{A numberic vector}

\item{na.rm}{Whether remove NA}

\item{zero.propagate}{666}
}
\description{
Calculate geometric mean
}
