% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mt_test}
\alias{mt_test}
\title{Test data}
\format{
A data frame with 61414 rows and 9 variables:
\describe{
  \item{sample}{TCGA sample barcode}
  \item{chromosome}{chromosome, chr}
  \item{position}{position for mutation}
  \item{MT_mean}{aggregated IC50}
  \item{exp}{expression of gene that has this mutation, TPM}
  \item{ccf_cn_assume}{CCF of mutation}
  \item{cancer_type}{TCGA cancer type}
  \item{ref}{ref allele}
  \item{alt}{alt allele}
}
}
\usage{
mt_test
}
\description{
A dataset containing mutations (SNV) of 1061 samples from TCGA
}
\keyword{datasets}
