% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cal_nes.R
\name{run_cal_nes}
\alias{run_cal_nes}
\title{run the function of calculating NES}
\usage{
run_cal_nes(
  cancer,
  calp = F,
  cal_type = "exp",
  mhc_type = "I",
  IC50_threshold,
  Rank_threshold
)
}
\arguments{
\item{cancer}{mutation file}

\item{calp}{bool value ,calculate p value or not}

\item{cal_type}{"exp" or "ccf"}

\item{mhc_type}{"I" or "II"}

\item{IC50_threshold}{threshold for MHC-I neoantigens}

\item{Rank_threshold}{threshold for MHC-II neoantigens}
}
\value{
a dataframe with es, nes, p_value
}
\description{
run the function of calculating NES
}
