% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_neutral.R
\name{simulation_neutral}
\alias{simulation_neutral}
\title{simulate neutral distribution of NES by cancer type}
\usage{
simulation_neutral(mutation_file)
}
\arguments{
\item{mutation_file}{At least contains columns of MT_mean(IC50) and cancer type for sampling and other columns for calculating NES}
}
\value{
cancer type neutral simulation
}
\description{
simulate neutral distribution of NES by cancer type
}
\examples{
neutral_simulation <- simulation_neutral(mt_test)
}
