% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_nes.R
\name{cales_t}
\alias{cales_t}
\title{calculate neoantigens NES for a sample}
\usage{
cales_t(
  data,
  barcode,
  calp = FALSE,
  cal_type = "exp",
  type = "I",
  sample_counts
)
}
\arguments{
\item{data}{Mutation file,folowing column are possible needed: MT_mean,exp,sample,chromosome,position,\%Rank_best_perL,ccf_cn_assume.}

\item{barcode}{Tumor sample barcode.}

\item{calp}{Bool value, calculate p value or not.}

\item{cal_type}{Should be "ccf" or "exp", calculate enrichment score in a aspect of Cancer Cell Fraction or Expression.}
}
\value{
es,nes,p
}
\description{
calculate neoantigens NES for a sample
}
