% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_KS.R
\name{plot_KS}
\alias{plot_KS}
\title{plot K-S like graph for a single sample}
\usage{
plot_KS(mutation_file, barcode, mhc_type = "I", IC50_threshold = 500)
}
\arguments{
\item{mutation_file}{Mutations with neoantigen prediction (IC50).}

\item{barcode}{TCGA sample barcode.}

\item{mhc_type}{MHC type ,can be I or II.}

\item{IC50_threshold}{Threshold which IC50 below it can be thought as neoantigen}
}
\description{
plot K-S like graph for a single sample
}
