% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pie.R
\name{plot_pie}
\alias{plot_pie}
\title{Plot pie}
\usage{
plot_pie(vector, expression, label)
}
\arguments{
\item{vector}{A numberic vector for plot}

\item{expression}{A Strings for calculate percent}

\item{label}{A character vector for display catalog}
}
\description{
Plot pie
}
